+++ pkg/services/ngalert/api/api.go
@@ -16,8 +16,8 @@ import (
 	"github.com/grafana/grafana/pkg/services/ngalert/metrics"
 	"github.com/grafana/grafana/pkg/services/ngalert/models"
 	"github.com/grafana/grafana/pkg/services/ngalert/notifier"
+	"github.com/grafana/grafana/pkg/services/ngalert/provisioning"
 	"github.com/grafana/grafana/pkg/services/ngalert/schedule"
-	"github.com/grafana/grafana/pkg/services/ngalert/services"
 	"github.com/grafana/grafana/pkg/services/ngalert/state"
 	"github.com/grafana/grafana/pkg/services/ngalert/store"
 	"github.com/grafana/grafana/pkg/services/quota"
@@ -65,7 +65,7 @@ type API struct {
 	ExpressionService    *expr.Service
 	QuotaService         *quota.QuotaService
 	Schedule             schedule.ScheduleService
-	TransactionManager   services.TransactionManager
+	TransactionManager   provisioning.TransactionManager
 	RuleStore            store.RuleStore
 	InstanceStore        store.InstanceStore
 	AlertingStore        AlertingStore
@@ -75,7 +75,7 @@ type API struct {
 	StateManager         *state.Manager
 	SecretsService       secrets.Service
 	AccessControl        accesscontrol.AccessControl
-	Policies             *services.NotificationPolicyService
+	Policies             *provisioning.NotificationPolicyService
 }
 
 // RegisterAPIEndpoints registers API handlers
