+++ pkg/services/ngalert/provisioning/notification_policies_test.go
@@ -1,4 +1,4 @@
-package services
+package provisioning
 
 import (
 	"context"
