+++ pkg/services/ngalert/provisioning/serialize.go
@@ -1,4 +1,4 @@
-package services
+package provisioning
 
 import (
 	"encoding/json"
