+++ pkg/services/ngalert/provisioning/testing.go
@@ -1,4 +1,4 @@
-package services
+package provisioning
 
 import (
 	"context"
